// apps/web/src/api/generated/model/organizationRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Starter API
 * OpenAPI spec version: 0.1.0
 */

export type OrganizationRole = typeof OrganizationRole[keyof typeof OrganizationRole];


export const OrganizationRole = {
  owner: 'owner',
  admin: 'admin',
  member: 'member',
  viewer: 'viewer',
} as const;