// apps/web/src/api/generated/model/organizationInvitation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Starter API
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationRole } from './organizationRole';

export interface OrganizationInvitation {
  id: string;
  organizationId: string;
  email: string;
  role: OrganizationRole;
  expiresAt: string;
  createdAt: string;
}